+++ src/main/java/nextstep/refactoring/ladder/concrete/LadderCreatorImpl.java
@@ -1,12 +1,13 @@
 package nextstep.refactoring.ladder.concrete;
 
-import static java.util.stream.Collectors.collectingAndThen;
-import static java.util.stream.Collectors.toList;
-
-import java.util.stream.Stream;
 import nextstep.refactoring.ladder.engine.Ladder;
 import nextstep.refactoring.ladder.engine.LadderCreator;
 import nextstep.refactoring.ladder.engine.LineCreator;
+
+import java.util.stream.Stream;
+
+import static java.util.stream.Collectors.collectingAndThen;
+import static java.util.stream.Collectors.toList;
 
 public class LadderCreatorImpl implements LadderCreator {
 
@@ -20,6 +21,6 @@ public class LadderCreatorImpl implements LadderCreator {
     public Ladder create(int height, int width) {
         return Stream.generate(() -> lineCreator.create(width))
                      .limit(height)
-                     .collect(collectingAndThen(toList(), lines -> new LadderImpl(lines, width)));
+                     .collect(collectingAndThen(toList(), lines -> new Ladder(lines, width)));
     }
 }
